fix: clean up example checks and formatting

+++ examples/01-renderers/ash_textures.rs
@@ -252,6 +252,7 @@ impl Drop for VulkanState {
 
 struct ImguiState {
     // Ensure registered textures are unregistered before the ImGui context is destroyed.
+    #[allow(dead_code)]
     registered_user_textures: Vec<dear_imgui_rs::RegisteredUserTexture>,
     context: Context,
     platform: WinitPlatform,
@@ -319,11 +320,7 @@ impl AppWindow {
         let tex_w: u32 = 128;
         let tex_h: u32 = 128;
         let mut img_tex = dear_imgui_rs::texture::TextureData::new();
-        img_tex.create(
-            dear_imgui_rs::texture::TextureFormat::RGBA32,
-            tex_w as i32,
-            tex_h as i32,
-        );
+        img_tex.create(dear_imgui_rs::texture::TextureFormat::RGBA32, tex_w, tex_h);
         let mut pixels = vec![0u8; (tex_w * tex_h * 4) as usize];
         for y in 0..tex_h {
             for x in 0..tex_w {
@@ -396,11 +393,7 @@ impl AppWindow {
         let (w, h) = rgba.dimensions();
 
         let mut tex = dear_imgui_rs::texture::TextureData::new();
-        tex.create(
-            dear_imgui_rs::texture::TextureFormat::RGBA32,
-            w as i32,
-            h as i32,
-        );
+        tex.create(dear_imgui_rs::texture::TextureFormat::RGBA32, w, h);
         tex.set_data(&rgba);
         Some(tex)
     }
